requirements.txt
PySide6>=6.6.0
picamera2>=0.3.18
PyYAML>=6.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
psutil>=5.9.0 